[UPD][ADD] report_csv, report_text_format_option
