+++ report_csv/__manifest__.py
@@ -6,10 +6,11 @@
     "author": "Creu Blanca, Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/reporting-engine",
     "category": "Reporting",
-    "version": "16.0.2.0.0",
+    "version": "16.0.1.0.0",
     "license": "AGPL-3",
-    "depends": ["base", "web", "report_format_option"],
+    "depends": ["base", "web"],
     "demo": ["demo/report.xml"],
+    "data": ["views/ir_actions_views.xml"],
     "assets": {
         "web.assets_backend": [
             "report_csv/static/src/js/report/qwebactionmanager.esm.js"
