+++ report_csv/readme/CONTRIBUTORS.rst
@@ -1,6 +1,3 @@
 * Enric Tobella <etobella@creublanca.es>
 * Jaime Arroyo <jaime.arroyo@creublanca.es>
 * Rattapong Chokmasermkul <rattapongc@ecosoft.co.th>
-* `Quartile <https://www.quartile.co>`__:
-
-  * Aung Ko Ko Lin
