+++ report_csv/views/ir_actions_views.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="act_report_xml_view" model="ir.ui.view">
+        <field name="name">ir.actions.report</field>
+        <field name="model">ir.actions.report</field>
+        <field name="inherit_id" ref="base.act_report_xml_view" />
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='report_type']" position="after">
+                <field
+                    name="encoding"
+                    attrs="{'invisible': [('report_type', '!=', 'csv')]}"
+                />
+                <field
+                    name="encode_error_handling"
+                    attrs="{'invisible': [('encoding', '=', False)]}"
+                />
+            </xpath>
+        </field>
+    </record>
+</odoo>
