+++ report_text_format_option/README.rst
@@ -1,13 +1,13 @@
-====================
-Report Format Option
-====================
+=========================
+Report Text Format Option
+=========================
 
 .. 
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:e3b9cc2a66291d03027c6100d82df8a6a0934f1802a392fe2ff1cad39557fe0d
+   !! source digest: sha256:37b5b67f110dcd3bf227357ada38dfcd3592d686401c059f8a337ed56881ffca
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 .. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
@@ -17,10 +17,10 @@ Report Format Option
     :target: http://www.gnu.org/licenses/lgpl-3.0-standalone.html
     :alt: License: LGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Freporting--engine-lightgray.png?logo=github
-    :target: https://github.com/OCA/reporting-engine/tree/16.0/report_format_option
+    :target: https://github.com/OCA/reporting-engine/tree/16.0/report_text_format_option
     :alt: OCA/reporting-engine
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/reporting-engine-16-0/reporting-engine-16-0-report_format_option
+    :target: https://translation.odoo-community.org/projects/reporting-engine-16-0/reporting-engine-16-0-report_text_format_option
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runboat-Try%20me-875A7B.png
     :target: https://runboat.odoo-community.org/builds?repo=OCA/reporting-engine&target_branch=16.0
@@ -29,7 +29,6 @@ Report Format Option
 |badge1| |badge2| |badge3| |badge4| |badge5|
 
 This is a technical module designed to add encoding fields to the ir.actions.report model and is applied to Text type reports.
-The module is expected to be a dependency of other reporting modules (e.g., report_csv).
 
 **Table of contents**
 
@@ -42,8 +41,8 @@ Configuration
 In case the exported report should be encoded in another system than UTF-8, following
 fields of the report record (*Settings > Technical > Reports*) should be populated accordingly.
 
-* Encoding: set an encoding system (such as cp932)
-* Encode Error Handling: select 'Ignore' or 'Replace' as necessary.
+* Text Encoding: set an encoding system (such as cp932)
+* Text Encode Error Handling: select 'Ignore' or 'Replace' as necessary.
   * 'Ignore': in case of an encoding error, the problematic character will be removed from the exported file.
   * 'Replace': in case of an encoding error, the problematic character will be replaced with '?' symbol.
   * Leaving the field blank: in case of an encoding error, the report generation fails with an error message.
@@ -58,7 +57,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/reporting-engine/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/reporting-engine/issues/new?body=module:%20report_format_option%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/reporting-engine/issues/new?body=module:%20report_text_format_option%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -83,6 +82,6 @@ OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-This module is part of the `OCA/reporting-engine <https://github.com/OCA/reporting-engine/tree/16.0/report_format_option>`_ project on GitHub.
+This module is part of the `OCA/reporting-engine <https://github.com/OCA/reporting-engine/tree/16.0/report_text_format_option>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
