+++ report_text_format_option/i18n/it.po
@@ -1,0 +1,93 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* report_text_format_option
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 16.0\n"
+"Report-Msgid-Bugs-To: \n"
+"PO-Revision-Date: 2024-06-07 12:34+0000\n"
+"Last-Translator: mymage <stefano.consolaro@mymage.it>\n"
+"Language-Team: none\n"
+"Language: it\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: nplurals=2; plural=n != 1;\n"
+"X-Generator: Weblate 4.17\n"
+
+#. module: report_text_format_option
+#: model:ir.model.fields.selection,name:report_text_format_option.selection__ir_actions_report__line_ending__cr
+msgid "CR (\\r)"
+msgstr "CR (\\r)"
+
+#. module: report_text_format_option
+#: model:ir.model.fields.selection,name:report_text_format_option.selection__ir_actions_report__line_ending__crlf
+msgid "CRLF (\\r\\n)"
+msgstr ""
+"CRLF (\\r\\n"
+")"
+
+#. module: report_text_format_option
+#: model:ir.actions.report,name:report_text_format_option.action_report_demo
+msgid "Demo Text Report"
+msgstr "Resoconto testo demo"
+
+#. module: report_text_format_option
+#: model:ir.model.fields,help:report_text_format_option.field_ir_actions_report__text_encoding
+msgid "Encoding to be applied to the generated Text file. e.g. cp932"
+msgstr "Codifica da applicare al file testo generato. Es. cp932"
+
+#. module: report_text_format_option
+#: model_terms:ir.ui.view,arch_db:report_text_format_option.demo_report_template
+msgid "Hello,"
+msgstr "Salve,"
+
+#. module: report_text_format_option
+#: model:ir.model.fields,help:report_text_format_option.field_ir_actions_report__text_encode_error_handling
+msgid ""
+"If nothing is selected, text export will fail with an error message when "
+"there is a character that fail to be encoded."
+msgstr ""
+"Se non viene selezionato nulla, l'esportazione del testo fallirà con un "
+"messaggio di errore quando è presente un carattere che non può essere "
+"codificato."
+
+#. module: report_text_format_option
+#: model:ir.model.fields.selection,name:report_text_format_option.selection__ir_actions_report__text_encode_error_handling__ignore
+msgid "Ignore"
+msgstr "Ignora"
+
+#. module: report_text_format_option
+#: model:ir.model.fields,field_description:report_text_format_option.field_ir_actions_report__line_ending
+msgid "Line Ending"
+msgstr "Fine riga"
+
+#. module: report_text_format_option
+#: model:ir.model.fields.selection,name:report_text_format_option.selection__ir_actions_report__text_encode_error_handling__replace
+msgid "Replace"
+msgstr "Sostituisce"
+
+#. module: report_text_format_option
+#: model:ir.model,name:report_text_format_option.model_ir_actions_report
+msgid "Report Action"
+msgstr "Azione resoconto"
+
+#. module: report_text_format_option
+#: model:ir.model.fields,help:report_text_format_option.field_ir_actions_report__line_ending
+msgid ""
+"Select the type of line ending in case the report needs to be output with "
+"other line ending than 'LF'."
+msgstr ""
+"Selezionare il tipo di fine riga nel caso il resoconto richieda di essere "
+"emesso con fine riga diverso da 'LF'."
+
+#. module: report_text_format_option
+#: model:ir.model.fields,field_description:report_text_format_option.field_ir_actions_report__text_encode_error_handling
+msgid "Text Encode Error Handling"
+msgstr "Gestione errore codifica testo"
+
+#. module: report_text_format_option
+#: model:ir.model.fields,field_description:report_text_format_option.field_ir_actions_report__text_encoding
+msgid "Text Encoding"
+msgstr "Codifica testo"
