+++ report_text_format_option/i18n/report_text_format_option.pot
@@ -1,0 +1,83 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* report_text_format_option
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 16.0\n"
+"Report-Msgid-Bugs-To: \n"
+"Last-Translator: \n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: report_text_format_option
+#: model:ir.model.fields.selection,name:report_text_format_option.selection__ir_actions_report__line_ending__cr
+msgid "CR (\\r)"
+msgstr ""
+
+#. module: report_text_format_option
+#: model:ir.model.fields.selection,name:report_text_format_option.selection__ir_actions_report__line_ending__crlf
+msgid "CRLF (\\r\\n)"
+msgstr ""
+
+#. module: report_text_format_option
+#: model:ir.actions.report,name:report_text_format_option.action_report_demo
+msgid "Demo Text Report"
+msgstr ""
+
+#. module: report_text_format_option
+#: model:ir.model.fields,help:report_text_format_option.field_ir_actions_report__text_encoding
+msgid "Encoding to be applied to the generated Text file. e.g. cp932"
+msgstr ""
+
+#. module: report_text_format_option
+#: model_terms:ir.ui.view,arch_db:report_text_format_option.demo_report_template
+msgid "Hello,"
+msgstr ""
+
+#. module: report_text_format_option
+#: model:ir.model.fields,help:report_text_format_option.field_ir_actions_report__text_encode_error_handling
+msgid ""
+"If nothing is selected, text export will fail with an error message when "
+"there is a character that fail to be encoded."
+msgstr ""
+
+#. module: report_text_format_option
+#: model:ir.model.fields.selection,name:report_text_format_option.selection__ir_actions_report__text_encode_error_handling__ignore
+msgid "Ignore"
+msgstr ""
+
+#. module: report_text_format_option
+#: model:ir.model.fields,field_description:report_text_format_option.field_ir_actions_report__line_ending
+msgid "Line Ending"
+msgstr ""
+
+#. module: report_text_format_option
+#: model:ir.model.fields.selection,name:report_text_format_option.selection__ir_actions_report__text_encode_error_handling__replace
+msgid "Replace"
+msgstr ""
+
+#. module: report_text_format_option
+#: model:ir.model,name:report_text_format_option.model_ir_actions_report
+msgid "Report Action"
+msgstr ""
+
+#. module: report_text_format_option
+#: model:ir.model.fields,help:report_text_format_option.field_ir_actions_report__line_ending
+msgid ""
+"Select the type of line ending in case the report needs to be output with "
+"other line ending than 'LF'."
+msgstr ""
+
+#. module: report_text_format_option
+#: model:ir.model.fields,field_description:report_text_format_option.field_ir_actions_report__text_encode_error_handling
+msgid "Text Encode Error Handling"
+msgstr ""
+
+#. module: report_text_format_option
+#: model:ir.model.fields,field_description:report_text_format_option.field_ir_actions_report__text_encoding
+msgid "Text Encoding"
+msgstr ""
