+++ report_text_format_option/readme/DESCRIPTION.rst
@@ -1,2 +1,1 @@
 This is a technical module designed to add encoding fields to the ir.actions.report model and is applied to Text type reports.
-The module is expected to be a dependency of other reporting modules (e.g., report_csv).
