+++ report_text_format_option/views/ir_actions_report_views.xml
@@ -6,19 +6,17 @@
         <field name="inherit_id" ref="base.act_report_xml_view" />
         <field name="arch" type="xml">
             <xpath expr="//field[@name='report_type']" position="after">
-                <field name="show_encoding" invisible="1" />
                 <field
-                    name="encoding"
-                    attrs="{'invisible': [('show_encoding', '=', False)]}"
+                    name="text_encoding"
+                    attrs="{'invisible': [('report_type', '!=', 'qweb-text')]}"
                 />
                 <field
-                    name="encode_error_handling"
-                    attrs="{'invisible': [('encoding', '=', False)]}"
+                    name="text_encode_error_handling"
+                    attrs="{'invisible': [('text_encoding', '=', False)]}"
                 />
-                <field name="show_line_ending" invisible="1" />
                 <field
                     name="line_ending"
-                    attrs="{'invisible': [('show_line_ending', '=', False)]}"
+                    attrs="{'invisible': [('report_type', '!=', 'qweb-text')]}"
                 />
             </xpath>
         </field>
