--- setup/report_format_option/odoo/addons/report_format_option
@@ -1,1 +1,0 @@
-../../../../report_format_option
