+++ setup/report_text_format_option/odoo/addons/report_text_format_option
@@ -1,0 +1,1 @@
+../../../../report_text_format_option
